refactor: split code-generation agent types
